requirements.txt
Django 
mysqlclient
djangorestframework
